refactor: update user menu
